// planning-idsdp-rms-frontend/src/app/main/modules/training-institute/models/agreement-parties.model.ts
export class AgreementPartiesModel {
    firstPartyAddress: String;
    firstPartyName: String;
    firstPartySurname: String;
    secondPartyAddress: String;
    secondPartyEmail: String;
    secondPartyName: String;
    secondPartyNidNo: String;
    secondPartyPhoneNo: String;
    secondPartySurname: String;
}